+++ Tests/CrispyWaffle.Tests/Serialization/TestObjects_/SampleJsonClass.cs
@@ -1,0 +1,120 @@
+﻿// ***********************************************************************
+// Assembly         : CrispyWaffle.Tests
+// Author           : Guilherme Branco Stracini
+// Created          : 09-05-2020
+//
+// Last Modified By : Guilherme Branco Stracini
+// Last Modified On : 09-05-2020
+// ***********************************************************************
+// <copyright file="SampleJsonClass.cs" company="Guilherme Branco Stracini ME">
+//     Copyright (c) Guilherme Branco Stracini ME. All rights reserved.
+// </copyright>
+// <summary></summary>
+// ***********************************************************************
+using CrispyWaffle.Serialization;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace CrispyWaffle.Tests.Serialization
+{
+    /// <summary>
+    /// Class SampleJsonClass.
+    /// Implements the <see cref="System.IEquatable{SampleJsonClass}" />
+    /// </summary>
+    /// <seealso cref="System.IEquatable{SampleJsonClass}" />
+    [Serializer(SerializerFormat.JSON)]
+    public class SampleJsonClass : IEquatable<SampleJsonClass>
+    {
+        #region Equality members
+
+        /// <summary>Indicates whether the current object is equal to another object of the same type.</summary>
+        /// <param name="other">An object to compare with this object.</param>
+        /// <returns>
+        /// <see langword="true" /> if the current object is equal to the <paramref name="other" /> parameter; otherwise, <see langword="false" />.</returns>
+        public bool Equals(SampleJsonClass other)
+        {
+            if (ReferenceEquals(null, other))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
+            return Id.Equals(other.Id) && Date.Equals(other.Date) && ListStrong.SequenceEqual(other.ListStrong);
+        }
+
+        /// <summary>Determines whether the specified object is equal to the current object.</summary>
+        /// <param name="obj">The object to compare with the current object.</param>
+        /// <returns>
+        /// <see langword="true" /> if the specified object  is equal to the current object; otherwise, <see langword="false" />.</returns>
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj))
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, obj))
+            {
+                return true;
+            }
+
+            if (obj.GetType() != this.GetType())
+            {
+                return false;
+            }
+
+            return Equals((SampleJsonClass)obj);
+        }
+
+        /// <summary>Serves as the default hash function.</summary>
+        /// <returns>A hash code for the current object.</returns>
+        public override int GetHashCode()
+        {
+            return HashCode.Combine(Id, Date, ListStrong);
+        }
+
+        /// <summary>Returns a value that indicates whether the values of two <see cref="T:CrispyWaffle.Tests.Serialization.SampleJsonClass" /> objects are equal.</summary>
+        /// <param name="left">The first value to compare.</param>
+        /// <param name="right">The second value to compare.</param>
+        /// <returns>true if the <paramref name="left" /> and <paramref name="right" /> parameters have the same value; otherwise, false.</returns>
+        public static bool operator ==(SampleJsonClass left, SampleJsonClass right)
+        {
+            return Equals(left, right);
+        }
+
+        /// <summary>Returns a value that indicates whether two <see cref="T:CrispyWaffle.Tests.Serialization.SampleJsonClass" /> objects have different values.</summary>
+        /// <param name="left">The first value to compare.</param>
+        /// <param name="right">The second value to compare.</param>
+        /// <returns>true if <paramref name="left" /> and <paramref name="right" /> are not equal; otherwise, false.</returns>
+        public static bool operator !=(SampleJsonClass left, SampleJsonClass right)
+        {
+            return !Equals(left, right);
+        }
+
+        #endregion
+
+        /// <summary>
+        /// Gets or sets the identifier.
+        /// </summary>
+        /// <value>The identifier.</value>
+        public Guid Id { get; set; }
+
+        /// <summary>
+        /// Gets or sets the date.
+        /// </summary>
+        /// <value>The date.</value>
+        public DateTime Date { get; set; }
+
+        /// <summary>
+        /// Gets or sets the list strong.
+        /// </summary>
+        /// <value>The list strong.</value>
+        public List<StrongTypingClass> ListStrong { get; set; }
+
+    }
+}